requirements.txt
git+https://github.com/Rapptz/discord.py.git@master
PyYAML~=6.0